Enable display of collection spec on details page
